+++ server/src/com/bitdreamit/connect/plugins/transmission/astm/server/ASTME1381TransmissionModePlugin.java
@@ -9,6 +9,7 @@ import com.mirth.connect.donkey.server.message.StreamHandler;
 import com.mirth.connect.donkey.server.message.batch.BatchStreamReader;
 import com.mirth.connect.model.transmission.TransmissionModeProperties;
 import com.mirth.connect.plugins.TransmissionModeProvider;
+import com.mirth.connect.model.converters.ObjectXMLSerializer;
 
 /**
  * Server-side transmission-mode plugin for ASTM E1381-02.
@@ -22,6 +23,29 @@ public class ASTME1381TransmissionModePlugin extends TransmissionModeProvider {
 
     public ASTME1381TransmissionModePlugin() {
         super();
+        whiteListPropertiesClass();
+    }
+
+    /**
+     * Mirth Connect 4.x deserializes channel XML through a security whitelist.
+     * The transmission-mode properties class embedded in channel exports must be
+     * allowed explicitly, otherwise the channel is stored as an invalid
+     * placeholder ("Verify all required extensions are loaded correctly").
+     */
+    private void whiteListPropertiesClass() {
+        try {
+            ObjectXMLSerializer serializer = ObjectXMLSerializer.getInstance();
+            java.lang.reflect.Method method = serializer.getClass()
+                    .getMethod("allowTypes", java.util.List.class, java.util.List.class, java.util.List.class);
+            java.util.List<String> allowedTypes = new java.util.ArrayList<String>();
+            allowedTypes.add(ASTME1381TransmissionModeProperties.class.getCanonicalName());
+            method.invoke(serializer, allowedTypes, null, null);
+            System.out.println("ASTM E1381 transmission mode properties whitelisted in XStream");
+        } catch (NoSuchMethodException e) {
+            System.out.println("Old Mirth Connect version, ASTM E1381 whitelist not required");
+        } catch (Exception e) {
+            System.err.println("Error whitelisting ASTM E1381 transmission mode properties: " + e);
+        }
     }
 
     @Override